+++ assets/test.lua
@@ -2,6 +2,7 @@
 local dapi = require("dapi")
 local dapi_os = require("dapi_os")
 local dapi_io = require("dapi_io")
+local dapi_http = require("dapi_http")
 
 -- TEST Dapi
 dapi.greet("ne")
@@ -44,3 +45,7 @@ local lines = dapi_io.read_line("README.md", 5)  -- Nur 5 Zeilen
 for i, line in ipairs(lines) do
     print(i, line)
 end
+
+print("Starting HTTP Server")
+local server = dapi_http.start_static_server("/", 8080)
+print("Server started")