+++ src/api/mod.rs
@@ -1,3 +1,4 @@
 pub mod base;
 pub mod io;
 pub mod os;
+pub mod http;